Move PVT callback math to hardware floats; dedup RMC toDegMin
The M4F FPU is single-precision - every double op is a software call.
The 25 Hz PVT callback did 7 software-double divides per frame; the
non-coordinate fields (altitude, speed, HDOP, heading, horizontal
accuracy) need nowhere near double precision, so GpsData carries them
as float and the callback computes them with reciprocal-constant
multiplies. lat/lng stay double: 1e-7 degree resolution over +/-180
needs ~2^31 steps, beyond float's 24-bit mantissa. All consumers take
these into double contexts (CourseManager, gps_validation, dtostrf,
Print) where float promotes losslessly.

The camera RMC builder computed toDegMin(latitude) three times per
10 Hz frame (ternary with repeated calls); cache it once. Output is
byte-identical by construction. The full double->fixed-point conversion
of the builder was evaluated and deliberately dropped: the digit chain
cannot be PROVEN byte-identical across the whole input domain
(representation-error boundary cases at the 4th minute decimal), and
the saving is ~4 double ops at 10 Hz - not worth risk on a golden-
pinned wire protocol.

Validation: insta360 golden-byte tests pass unmodified; both sim lap
oracles reproduce the hardware fixture to the exact millisecond.

+++ BirdsEye/BirdsEye.ino
@@ -307,14 +307,18 @@ SFE_UBLOX_GNSS_SERIAL myGNSS;
 bool gpsInitialized = false;  // Safety flag - true only after successful GPS init
 
 // Cached PVT data — updated by onPVTReceived() callback from checkCallbacks()
+// lat/lng stay double: 1e-7 deg resolution over ±180° needs ~2^31 steps,
+// beyond float's 24-bit mantissa. Everything else is float — well within
+// 7 significant digits, and double math is SOFTWARE-emulated on the
+// M4F's single-precision FPU (consumers that take double promote fine).
 struct GpsData {
   double latitudeDegrees;
   double longitudeDegrees;
-  double altitude;       // meters
-  double speed;          // knots (for DovesLapTimer compatibility)
-  double HDOP;
-  double heading;            // degrees (0-360), heading of motion
-  double horizontalAccuracy; // meters, horizontal accuracy estimate
+  float altitude;       // meters
+  float speed;          // knots (for DovesLapTimer compatibility)
+  float HDOP;
+  float heading;            // degrees (0-360), heading of motion
+  float horizontalAccuracy; // meters, horizontal accuracy estimate
   int satellites;
   bool fix;
   bool timeValid;        // true only when the module reports validDate+validTime+fullyResolved
